+++ uui/src/lib/List/List.types.d.ts
@@ -1,0 +1,4 @@
+import type { BoxProps } from '../Box/box.types';
+export interface ListProps<T extends EventTarget> extends BoxProps<T> {
+  disablePadding?: boolean;
+}