+++ uui/src/lib/List/ListSubHeader.types.d.ts
@@ -1,0 +1,1 @@
+export type ListSubHeaderProps = svelte.JSX.HTMLAttributes<HTMLHeadingElement>;