+++ uui/src/lib/index.ts
@@ -2,6 +2,9 @@
 import Box from './Box/box.svelte';
 import Chip from './Chip/chip.svelte';
 import Step from './Stepper/Step.svelte';
+import List from './List/List.svelte';
+import Paper from './Paper/Paper.svelte';
+import Overlay from './Paper/Overlay.svelte';
 import Theme from './Theme/theme.svelte';
 import Button from './Button/Button.svelte';
 import Dialog from './Dialog/Dialog.svelte';
@@ -9,34 +12,47 @@ import Stepper from './Stepper/Stepper.svelte';
 import ChipIcon from './Chip/chip-icon.svelte';
 import StepIcon from './Stepper/StepIcon.svelte';
 import TextField from './TextField/TextField.svelte';
+import ListItem from './List/ListItem.svelte';
+import ListGroup from './List/ListGroup.svelte';
 import Accordion from './Accordion/accordion.svelte';
 import ButtonIcon from './Button/ButtonIcon.svelte';
 import IconButton from './IconButton/IconButton.svelte';
 import Typography from './Typography/typography.svelte';
 import DialogTitle from './Dialog/DialogTitle.svelte';
-import TextFieldIcon from './TextField/TextFieldIcon.svelte';
+import ListSubHeader from './List/ListSubHeader.svelte';
 import DialogActions from './Dialog/DialogActions.svelte';
+import TextFieldIcon from './TextField/TextFieldIcon.svelte';
 import AccordionGroup from './Accordion/accordion-group.svelte';
+import ListItemCheckbox from './List/ListItemCheckbox.svelte';
+import ListItemAdornment from './List/ListItemAdornment.svelte';
 
 export {
   Box,
   Chip,
   Step,
+  List,
+  Paper,
+  Overlay,
   Theme,
   Button,
   Dialog,
   Stepper,
   ChipIcon,
   StepIcon,
+  ListItem,
+  ListGroup,
   TextField,
   Accordion,
   ButtonIcon,
   IconButton,
   Typography,
   DialogTitle,
+  ListSubHeader,
   TextFieldIcon,
   DialogActions,
-  AccordionGroup
+  AccordionGroup,
+  ListItemCheckbox,
+  ListItemAdornment
 };
 
 // Export Utility Functions here
@@ -47,6 +63,9 @@ export { UUI_PALETTES } from './shared/theme/palette/default-palettes';
 export { UUI_BREAKPOINTS } from './shared/theme/breakpoints/default-breakpoints';
 export { BREAKPOINT_CONTEXT } from './shared/theme/breakpoints/breakpoint-context';
 export { UUI_TYPOGRAPHY } from './shared/theme/typography-config/default-typography-config';
+
+// Export Stores here
+export { currentTheme } from './shared/theme/store/theme-store';
 
 // Re-export your types here.
 export type {
@@ -96,8 +115,20 @@ export type {
 export type { TextFieldProps, TextFieldIconProps } from './TextField/TextField.types';
 export type { AccordionProps, AccordionGroupProps } from './Accordion/accordion.types';
 export type { ThemeColors } from './shared/theme/default-colors/colors.types';
+export type { ComponentAsProp } from './shared/types/ComponentAsProp';
+export type { ListGroupProps } from './List/ListGroup.types';
+export type { ListSubHeaderProps } from './List/ListSubHeader.types';
+export type { ListProps } from './List/List.types';
+export type {
+  ListItemProps,
+  ListItemSlots,
+  ListItemSlotProps,
+  ListItemAdornmentProps
+} from './List/ListItem.types';
 export type {
   Breakpoints,
   BreakpointKey,
   BreakpointValue
 } from './shared/theme/breakpoints/breakpoints.types';
+export type { PaperProps } from './Paper/Paper.types';
+export type { OverlayProps } from './Paper/Overlay.types';